Support GA4 and Universal Analytics double tracking. #16
